added train errors, Expanded the SVM models and added a comment on plans.
Added a function that computes train accuracy. Added two SVM models looking at how the SVM performs with few parameters. Added a comments at end coming to the conclusion that we need more data and plans for what this data could be and how I could go about getting it.